Adjust operator precedence
As discussed on the libraries list, this is a bit of a minefield.  I
want cons and snoc operators and patterns to have different precedence
in case someone wants to write something like `x :< as :> y`. I wish I
could give one operator precedence 5 and the other
4.5 or 5.5, but we don't have fractional precedence. Precedence
5 matches that of `:`, which is pleasantly higher precedence than
`Functor` and `Applicative` operators and lower precedence than
standard arithmetic operators. Running into arithmetic seems
particularly troublesome, so let's go with 4 and 5. There's no
good choice for which gets which, since queues want one thing
and stacks want the other. I'm going with `:<` having the higher
precedence to match `:`; nothing very principled, but that was
Henning Thielemann's instinct.

Closes #13